backend/requirements.txt
Django==4.2.4
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.0
oauthlib==3.2.2
psycopg2==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
django-silk==5.0.3
